Split massive CMap.h a bit
