+++ client/lobby/SelectionTab.cpp
@@ -37,6 +37,7 @@
 #include "../../lib/filesystem/Filesystem.h"
 #include "../../lib/mapping/CMapInfo.h"
 #include "../../lib/mapping/CMap.h"
+#include "../../lib/mapping/MapFormat.h"
 #include "../../lib/mapping/CCampaignHandler.h"
 #include "../../lib/serializer/Connection.h"
