+++ lib/mapping/CMapDefines.h
@@ -12,9 +12,13 @@
 
 VCMI_LIB_NAMESPACE_BEGIN
 
+#include "../ResourceSet.h"
+
 class TerrainType;
 class RiverType;
 class RoadType;
+class CGObjectInstance;
+class CGTownInstance;
 
 /// The map event is an event which e.g. gives or takes resources of a specific
 /// amount to/from players and can appear regularly or once a time.
